[refactor] remove EncryptionKey
